refactored paginated view and refactored paginated sliver list and grid views
