+++ lib/widgets/paginated_sliver_list_view.dart
@@ -4,18 +4,16 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../paginated_search.dart';
 import 'abstract_paginated_view.dart';
 
+// Helper class for showing items of a paginated search in a sliver list view
 class PaginatedSliverListView<T, F> extends AbstractPaginatedView<T,F> {
-
-  final SliverGridDelegate gridDelegate;
 
   const PaginatedSliverListView({
     super.key,
-    required this.gridDelegate,
     required super.paginatedController,
     required super.itemBuilder,
-    super.emptyBuilder,
-    super.errorBuilder,
     super.loadingBuilder,
+    super.errorBuilder,
+    super.emptyBuilder,
   });
 
   @override
@@ -23,8 +21,7 @@ class PaginatedSliverListView<T, F> extends AbstractPaginatedView<T,F> {
     if (items.isEmpty && emptyBuilder != null) {
       return emptyBuilder?.call(context) ?? const SliverToBoxAdapter();
     }
-    return SliverGrid(
-      gridDelegate: gridDelegate,
+    return SliverList(
       delegate: SliverChildBuilderDelegate(
             (context, index) => itemBuilder(items[index]),
         childCount: items.length,
